Correct bounds overrun in motion compensation (1fmv8)
1fmv8 does subpel resampling using a 6-tap filter that will extend
outside the bounds of the old OD_UMV_PADDING around the reference
buffers.  It's necessary to either increase the padding to take the
extra filter width into account, or special-case the resampling filter
in 1fmv8 the same way as it was special-cased in upsample8.

To avoid the extra code complexity, we opt for increasing the buffer
padding for now.  Before release, we should change to special-casing
(ala the "emulated edge MC" approach of libav/FFmpeg).

This patch also separates the conflated meanings of the padding and
UMV clamp; they are now separate constants.

Conflicts:
	src/decode.c
	src/encode.c